@@ -44,10 +44,10 @@ class TestDistAggregateRegressor:
             cnt += 1
         expected_predictions = np.array(
             [
-                0.3045717477798462,
-                1.1020143032073975,
-                0.6729293465614319,
-                0.8377850651741028,
+                0.44149839878082275,
+                0.4521682560443878,
+                0.35752978920936584,
+                0.8726186752319336,
             ]
         )
 
@@ -79,10 +79,10 @@ class TestDistAggregateRegressor:
             cnt += 1
         expected_predictions = np.array(
             [
-                1.0728034973144531,
-                -0.14423903822898865,
-                1.1754292249679565,
-                0.2284281849861145,
+                0.7171313762664795,
+                0.8493690490722656,
+                0.6237673759460449,
+                -0.03409786522388458,
             ]
         )
         assert np.allclose(y_pred, expected_predictions)
